Uploading revised encoding and adding facsimile
This is the file we will use to begin final spacing and zone boxing before converting to publication format. All edits should be kept up-to-date in this file, whichever working file they are made in, or those changes will not be included in the published version of the notebook. I have significantly revised the editor's encoding to fix encoding errors and make the file fully valid, including changing some attribute values and adding missing tags, moving misplaced tags, and other edits.